requirements.txt
Django>=5.0,<6.0
python-dotenv>=1.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3